Initialize variable declarations in QQuickFlickablePrivate::AxisData
Not all of them were initialized in the constructor, which was confusing
during debugging, even if there were no known bugs due to that.

Also get rid of reserved bits. We don't care if this private data
changes size from one Qt version to another; but it's best to try to
reduce the size over time. There was risk of growing because of
misunderstanding memory layout; and let's keep the mutable bitfield
variables separate, in case some compilers segregate the storage.

Pick-to: 6.8
Change-Id: I16c9348d20fa209a603dad2abf10337fcbe40c26
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>