Item/PathView: reset the currentItem before releasing it
releaseItem ends up in code paths that emit signals and might activate
bindings or handlers, potentially leading to recursion. Since the
released item is deleted in releaseItem, this results in a dangling
pointer and potential crashes.

To prevent that, reset the currentItem pointer before we release it.
Add a helper function that encapsulates that before calling the (virtual
in case of QQuickItemView) releaseItem method.

Pick-to: 6.8 6.7 6.5
Fixes: QTBUG-122998
Change-Id: I96f2ea35a774cf3dae09e25342e177449a3980c3
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>