QtQml: Fix handling of redirected qmldirs and their URIs
If the qmldir gets redirected via a "prefer" directive we
a, Should update the URI of the corresponding import from the new
   qmldir. The original URI may be some guess for the implicit import.
b, Load the non-plugin types only after fixing the URI

tst_qquicklistview2 had a module called "test" and another one called
"Test". Those were confused on case-insensitive file systems. The change
to the redirection handling would cause the URI to be updated after the
confusion and the wrong C++ types to get loaded. This is considered a
user error. We therefore rename one of the modules to avoid it.

Amends commit 05a9fa44378a96838d56b37eaa59223919d5a721

Pick-to: 6.8
Fixes: QTBUG-126504
Change-Id: If8e1789726cc76af3b5d87089e95bcc6f5f3a1e4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>