QQuickPopupWindow: inform the popup that it's hiding
The current implementation would just set popup->visible
to false when the window was hidden from e.g clicking
outside of it. But just setting visible to false will
not inform the popup that it's been hidden. And the
popup needs to know to e.g clear the currently
hovered menu bar item etc.

This patch will change the code, so that we instead
call popup->setVisible(false). Then the popup can do
whatever it needs to do, and at the same time, emit
the visibleChanged signal.

Pick-to: 6.8
Change-Id: I9d314f5afb4f99dce20f97ba9ab71341f3565ce1
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>