QQmlDelegateModel: fix delegates not being created in certain cases
When calling beginResetModel from the constructor of a
QAbstractItemModel subclass, QQmlDelegateModel has not yet connected
to modelAboutToBeReset. endResetModel is then called later, but after
837c2f18cd223707e7cedb213257b0158ea07146, we no longer connect to
modelReset, so we never caught this and delegates wouldn't be visible.

This patch ensures that we connect to modelReset and call
QQmlDelegateModel::handleModelReset().

Fixes: QTBUG-125053
Pick-to: 6.7 6.5
Change-Id: Ib9c179420053a8733b505f3d8d51c712df28a348
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 6561344dd2d1ba69abe6edec4fe340b256da9e13)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>