Update generated assets in weather forecast example
This have been regenerated with an updated svgtoqml. One
change was needed in the example: The generator would
previously added a subitem inside the main item which was
transformed to fit the item to the requested size:

    Item {
        implicitWidth: foo; implicitHeight: bar
        Item {
            transform: ...
            Shape { ... }
        }
    }

This was since removed, likely because it's dead weight
(and removes the possibility of overriding it), so now
the transform is set directly on the generated root
object.

But the weather forecast example had made the map labels
children of the generated item, which means the scale was
now also applied to them. This was easy to fix in the
example, by creating the extra item ourselves in the
case where it's needed.

This patch also takes the liberty of fixing some minor
whitespace issues etc.

Change-Id: Ic324815c71f990bb7e8e7caed659c14267b1777a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit a784238e81a15a9768855fdd6db28c4eac92e01e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>