QQuickMenu: ensure we pick the correct menu item on release
The current implementation would use eventPoint->position()
to find the menu item under the pointer. But this position
was relative to the popup window, and not the ListView. This
would result in a bug where the wrong menu item was sometimes
triggered.

Moreover, the function would drop finding the menu item under
the pointer if a menu item was already made current. But this
meant that we would trigger the last hovered menu item, even if
the pointer was actually released outside the menu.

This patch will therefore make sure that we map the event
position to the ListView, and always pick the menu item under
the mouse, rather than relying on currentIndex.

Change-Id: Iec3504e7f35bd490d11aa196403f4f69735c42fb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit c45fd4a651ba325f6b8fab774a30321bd7475e94)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>