QQuickPopupWindow: close all menus after a press'n'drag
As it stood, if the user did a press'n'drag, but decided to
release the pointer outside any of the open menus, the menus
would stay open. The correct way is that all the menus should
always close after a press'n'drag.

This patch will therefore make sure that we close all menus
after a press'n'drag. And, since it was a bit confusing to
work with the code as it stood, the opportunity was taken to
refactor it a bit, so that we talk about "targetMenu" and
"targetMenuBar", rather than "targetPopupWindow" and
"targetWindow".

Pick-to: 6.8
Change-Id: Iec2266e171dae3642f43fd89a05dbab21be40433
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>