QQuickPopupWindow: close all menus after a press'n'drag
As it stood, if the user did a press'n'drag, but decided to
release the pointer outside any of the open menus, the menus
would stay open. The correct way is that all the menus should
always close after a press'n'drag.

This patch will therefore make sure that we close all menus
after a press'n'drag. And, since it was a bit confusing to
work with the code as it stood, the opportunity was taken to
refactor it a bit, so that we talk about "targetMenu" and
"targetMenuBar", rather than "targetPopupWindow" and
"targetWindow".

Change-Id: Iec2266e171dae3642f43fd89a05dbab21be40433
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 5c83765c3f0e9cb96d3e0e9c431a6bfe3fe7e0be)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>