Reset hover state for the items within the quick popup window
The hovered state of the item within the quick popup has not been
reset when another modal popup is placed over it. This is because,
after patch 48da34f3d797f58b3eda9ed0a98aa1776d75d946, the overlay
filters the hover events of the quick controls which are modally
blocked.

But this can create an issue if the child item of the pop is
already in a hovered state and there is another modal popup placed
over it, as any further hover events (including the HoverLeave of
the control item) will be forwarded and handled by the modal popup
which doesn't allow the child item to resets its hovered state.

This patch allows the child item of the popup to handle hover
release event when it's already in the hovered state.

Fixes: QTBUG-126626
Pick-to: 6.8 6.7 6.5 6.2
Change-Id: Ibf7f5ad8653a8c0df9968114ab6b8228acba9d6e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>