Popup: let the style set the default popup type
Rather than hard-coding from c++ which popup type
to use, leave it to the style to set the preferred
type. This let's us e.g enable native menus for the
macOS style, but keep the menu implementations for
the other (perhaps 3rd party) styles, even when
running on macOS. For example the Material and Imagine
styles have menus that matches the other controls.

Since the style will decide the popup type, we can no
longer support setting the popupType to "undefined".
But being able to change the popup type back and forth
at runtime (and resetting it) is a very low-priority
use case, and can be worked around by other means (the
app can store the initial value before changing it), if
needed.

Pick-to: 6.8
Change-Id: I95575bd25b313a4141007c6f16c4cda0482c573b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>