QQDeliveryAgent: check every touchpoint for double-tap violations
In a quick sequence in which one item is tapped, another is tapped, and
the first is tapped again: if the first item requires synth-mouse events
but the second item accepts touch events, do not give the first item a
MouseButtonDblClick event. The intervening touch event in a completely
different place should be considered to interrupt the synthetic
double-click. In fact we don't even care if it's two different items:
if the distance exceeds touchDoubleTapDistance or the time delta exceeds
mouseDoubleClickInterval, it's not a double-tap.

Refactor the constraint-checking functions, since we're now checking
constraints in multiple places. And fix an old mistake in
tst_touchmouse while we're at it.

QQDAPrivate::isWithinDoubleClickInterval() is meant to be a reusable
utility function; use it in QQuickAbstractButtonPrivate::handleRelease()
just to prove that it's possible. There are other places where we're
checking qApp->styleHints()->mouseDoubleClickInterval() which can
potentially be replaced later.

Fixes: QTBUG-116442
Pick-to: 6.5 6.7 6.8
Change-Id: I2599cfca1ac5ff61ba6124df86d7a9bfe6a836ee
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>