Fix flaky test tst_qquicklistview::multipleTransitions
The test assumed that an item inserted in the beginning of the ListView
will always end up at position 0 (x or y).

This is wrong assumption, since the ListView can choose to put the new
items above/left of the existing items without having to adjust the
coordinates of all of the existing items, and then accordingly it will
adjust the origin(X,Y), ensuring that the visual rendering will
be the same for both cases.

The documentation of Flickable::originY explicitly mentions this
(https://doc.qt.io/qt-6/qml-qtquick-flickable.html#originY-prop):

 "originY : real
  [...]
  This is usually (0,0), however ListView and GridView may have an
  arbitrary origin due to delegate size variation, or item
  insertion/removal outside the visible region."

This can also be reproduced without any transitions, so assuming that
the y coordinate of the first item is always 0 should therefore strictly
not be part of a test intended for transitions.

Pick-to: 6.7 6.8
Fixes: QTBUG-123894
Change-Id: Idc63cd385ed62936d9e45e31fcf3f725ee4405c4
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>