qmllint: avoid stale typeresolver pointer
It's not accessible via public API, but it risks wrecking havoc on
internal plugins, and would also likely cause issues if we were to
expose the global JS object to QQmlSA plugins.

The current fix is rather a minimal workaround, we should revisit
whether codegen should really take ownership of the typeResolver.

Pick-to: 6.8
Change-Id: I6ca6e78527ce886155cfc0a21038348da2982c03
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>