FluentWinUI3 Style: Use platform palette on Windows11+ platforms
When on Windows11+ platforms, use system colors as reported by the
platform theme.
Also don't use BrightText color role for pressed button text in the
FluentWinUI3 style, but hardcode this color. We do the same in the
windows11 widgets style. This is because the windows theme uses the
BrightText role for a different purpose.

Task-number: QTBUG-125279
Change-Id: I76c710145e0f9fdfbfe058300ba4477020bd62e9
Pick-to: 6.8
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>