QuickShapes: Unify the renderers' handling of 0 strokeWidth
A stroke of zero width should not be visible: that seems to be the
intuitively expected behavior, and corresponds to what SVG
does. However, in Qt Quick shapes, only the curve renderer backend
behaved like that. The geometry and software renderers would behave
differently, and different to each other. This commit standardizes on
the SVG behavior.

Fixes: QTBUG-126819
Pick-to: 6.8
Change-Id: I63410a732066facfad0a5f6e0565e2e80becb397
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>