Avoid infinite loop in ListView/GridView snapOne/snapTo tests
I'm not sure if any actual CI failures resulted from failure to
terminate the do..while loops, but it seems likely, since a recent patch
seems to provoke that. It's better to fail after a reasonable number of
loops, rather than letting it go and hoping for a timeout. In practice
it seems 2 rounds is enough, so let's try 4 as the limit.

While we're at it, round the grid.contentX/contentY values that are
displayed. Intermediate values have always been shown with way too many
decimal places, which interfered with whatever value it has to display
the numbers.

Pick-to: 6.8
Task-number: QTBUG-33017
Task-number: QTBUG-88644
Change-Id: I2fc4843071276b3af544c2506eb152698542cd06
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>