Compiler: Weaken god function populateReaderLocations
The function populateReaderLocations does too much and is very hard to
understand. This patch breaks up its functionality into several pieces.
populateReaderLocations keeps the code related to what its name
suggests it does: determining, for each instruction, what it reads and
writes. The dead store elimination part is extracted into a separate
function.

The TODO for the merging of conversions types using QSet::unite() was
also removed. This function was recently optimized in
92acc94fa98d19929d28feb5d543acf8c50c2290 and can be used again.

Change-Id: Iad0870e08d09b47acbc51afa96e582f95e61b18f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>