sg: Opt out of compute-related resource tracking
Copy the flag from Quick 3D, which used to set this for quite some time
now as a microoptimization. It is not great technically and may need to
be removed in the future once compute is used in some form in
combination with textures or buffers that are also used in the Qt Quick
render pass.

Task-number: QTBUG-125087
Pick-to: 6.8 6.7 6.6 6.5
Change-Id: Ie2d4d6a784580e669fe0fc0b843519b4826efa34
Reviewed-by: Andy Nichols <andy.nichols@qt.io>