Quick Dialogs: Opt out of QTP0004
This silences the CMake warning about using the old behavior of the
policy. We however don't want the additional qmldirs, as that seem to
cause name clashes (we have QML types with the same names as C++ types
in QtQuick.Dialogs and QtQuick.Dialogs.quickimpl; consequently the
import order really matters).

Opting in to the policy can be addressed in the refactoring that should
be done as part of QTBUG-104091.

Task-number: QTBUG-126201
Pick-to: 6.8
Change-Id: I461c675435df070db17b4cb930fae0de9a7edf3d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>