qmlls: don't resolve component types as attached types
Don't resolve component names as attached types. Instead, check whether
the component name has a field member access to a property name, that
is, whether the type name is followed by a dot and a lowercase name,
like in "MyApplicationWindow.header" where MyApplicationWindow is
an attached type because header starts with a lower case letter, or in
"MyApplicationWindow.Kitty" and "MyApplicationWindow.Hello.World" where
MyApplicationWindow is a component because "Hello" and "World" start
with an upper-case letter, for example.

With the improved handling in resolveExpressionType, fix the
findDefinitionFromLocation logic to also work on attached types.

Implement support for qualified module names in resolveExpressionType
and ignore unqualified modules names for now, so that component types
from qualified module names can be resolved correctly too. Collect the
used qualified module names in the qqmljsimportvisitor and pass them on
via the typeresolver to resolveExpressionType to recognize qualified
module names. Those do not have a qqmljsscope but still can be resolved
when used in a fieldmemberexpression.

Pick-to: 6.8
Fixes: QTBUG-126711
Change-Id: I274c644d25cd6f9728bc56f19ec96983ad217df8
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>