sg: Shortcut material ubuf update for gl
Only for OpenGL: there a QRhiBuffer with UniformBuffer usage is not
backed by a native OpenGL buffer at all. The ominuously named
beginFullDynamicBufferUpdateForCurrentFrame() just returns a pointer
into the underlying QByteArray (or similar) storage then. (only on GL,
only for uniform buffers) Knowing this, the batches' uniform data
can be memcopied directly into this area, instead of enqueing resource
updates that ultimately achieve the same.

Could be applicable to D3D11, to save a memcpy. However, it is not given
that, thinking of unmerged batches, separately memcopying lots of small
blocks (Qt Quick does not actually build up the final, full uniform
buffer contents itself) into a buffer's mapped (host visible) storage is
a good thing. (given the potentially different memory caching behavior
e.g.) So it is not done there.

To recognize the special case, we use nativeBuffers().slotCount == 0
which is only every true with the OpenGL and Null backends (indicating
that the QRhiBuffer has 0 real buffer resources associated), which
conveniently avoids the need to introduce backend-specific conditions.

Pick-to: 6.8 6.7 6.6 6.5
Task-number: QTBUG-125087
Change-Id: I88a9ef5eff6c52a88b2605bdab8cada86a18e98e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>