Fix popup windows emitting closed signal twice during exit transition
When a popup is closed, and the popup type is set to PopupWindow, the
exit transition is called twice, resulting in closed() signals emitting
twice.

There is a guard in QQuickPopup::setVisible that checks whether
visibility is always set to true. In the case it is an exit transition,
only once the transition has completed, will the visibility be false.
Therefore, the guard in QQuickPopup::setVisible that checks whether the
visibility is being set to the value it already is has to consider
whether an exit transition is taking place, to allow a popup to stop an
exit transition and start an enter transition.

What this logic does not consider, is that when destructing a popup
window, QQuickPopup::setVisibility(false) may be called several times
during the exit transition. To get around this, an extra check is needed
such that when visiblility is false and is being set to false, return.

Fixes: QTBUG-126858
Change-Id: I02f9ed52900c452972d01094d8ebc0936c16d617
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit d1c5a87f0cb1bb06aac3fb6ad911fc1a447f8f70)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>