QQmlTypeWrapper: Fix attached property lookup
When we made QQmlTypeWrapper usable as a constructor, we had to give it
Function as its proto-type. That had the unfortunate side-effect that a
lookup of an attached property would instead find the property from the
function prototype.
Avoid that issue for now by using the fallback getter for such lookups
(which will in turn call QQmlTypeWrapper::virtualGet which works
correctly).

Amends f62adc926f34cb28831ead07d716fedd8d6d6c7b

The special case for the attached property currently uses exactly the
same code as the catch-all fallback; however, it's intentionally kept
there to retain more symmetry with virtualGet. Moreover, follow-up
optimization of the lookup will look different.

Optimizing those lookups again is left as a future exercise.

Pick-to: 6.8
Task-number: QTBUG-126863
Change-Id: I8a29daef4195f4ee2ef4dc3c7af6eb1451800beb
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>