sg: Use the new buffer update shortcut for geometry with d3d11 and gl
If we know the (dynamic) vertex QRhiBuffer is backed by a single native
buffer object, we could use the
begin/EndFullDynamicBufferUpdateForCurrentFrame mechanism to update the
contents since the entire buffer contents is changing, saving a memcpy
in practice (and no need to enqueue a resource update via the standard
QRhi mechanism)

However, with OpenGL glMapBufferRange has "interesting" behavior
patterns on some platforms (like unexpectedly trippling the CPU load for
no apparent reason on some systems), and anyway all we want in the end
with GL is a glBufferSubData. So instead of the mapBuffer-like existing
begin/end functions, use a newly introduced helper function in
QRhiBuffer that, specifically for OpenGL, boils down to doing a
glBufferSubData, whereas it is equivalent to begin+memcpy+end everywhere
else.

Not applicable for backends for modern APIs.

The logic is based on querying resourceLimit(FramesInFlight) == 1. This
is convenient to avoid having to write backend type checks, although it
relies on undocumented internal knowledge (as FramesInFlight==1 may not
strictly imply and give guarantees on how a Dynamic QRhiBuffer is
implemented internally)

Task-number: QTBUG-125087
Change-Id: I317f9ff4bc03fadb6808b61cb053461b30ef1db7
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 1f29bc32dffef331d114e9901751f445d7bc2c89)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>