TreeViewDelegate: don't use palette settings in Material style
When the app's theme was different from the system theme, the
TreeViewDelegate didn't display properly as the text color was the same
as the background color.

The reason was that the TreeViewDelegate used the palette colors in the
Material style.

To fix this issue, the TreeViewDelegate should use the Material
settings rather than palette settings.

Fixes: QTBUG-125296
Pick-to: 6.7 6.5
Change-Id: I8c34511ed714fe709296df02873a7ebe873a4361
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 7bca438fcb4294deb4b002753baa3780a5928e46)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>