QQuickFolderListModel: ensure that beginResetModel isn't called twice
In the process of opening a FolderDialog, FolderListModel's folder
property is set twice: once upon component completion of
FolderDialog.qml, and shortly after if FolderDialog's currentFolder
property is set, due to the binding in FolderDialog.qml. This would
cause beginResetModel to be called twice, which invalidates the
assertion made in QQmlDelegateModel that this doesn't happen.

This patch ensures that beginResetModel is always followed by
endResetModel by tracking if we're in the middle of a reset.

This is (indirectly) tested by
tst_QQuickFolderDialogImpl::bindCurrentFolder.

Task-number: QTBUG-125053
Pick-to: 6.5 6.7 6.8
Change-Id: I4c22be173807187e56f4d528a465225b787e840a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>