Android: QML Embedding: Add Kotlin example showcasing QtAbstractListModel
This commit adds two projects, one for Android Studio and the other a
regular Qt QML project.

The Android Studio project provides two buttons which add and remove
rows, a frame for the embedded QML Component to be drawn on, and the
QtAbstractListModel which is passed to the QML Component.

The QtAbstractListModel contains a list of UUID strings which are
randomly generated at init and on request via the Android UI. This data
provided to QML has two properties: the UUID string and the row of the
requested data as an int.

The QML project simply has a ListView that uses the provided
QtAbstractListModel to draw elements that display the element data.

Task-number: QTBUG-125976
Change-Id: I315a85e6d0043c1ff3a15e12458c27f38f573e0b
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>
(cherry picked from commit 1b85df0a2eadfe33ada60e4f444feb8cc68a3e90)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>