DesignerSupport: Build on QQmlOpenMetaObject
If we assign a QQmlOpenMetaObject as QMetaObject to an object, we also
need to use it as the QDynamicMetaObjectData. Otherwise we get UB by
static_cast'ing to the wrong type. At that point, however, we can just
let QQmlDesignerMetaObject be a QQmlOpenMetaObject and save ourselves
quite a bit of hassle.

Some care need to be taken to ensure that we properly invalidate the
property cache in response to a createProperty call, just like we do in
createNewDynamicProperty.

As a side effect, we can re-use the property cache that
QQmlOpenMetaObject already maintains (when requested to do so) and
thereby prevent the property caches from going out of sync if some other
code makes the QQmlOpenMetaObject create a new property.

Original-patch-by: Ulf Hermann <ulf.hermann@qt.io>
Pick-to: 6.7
Fixes: QTBUG-126690
Change-Id: I2b00cd239820a6b8c7a17206ecafa5562302ed9e
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>
(cherry picked from commit 073243256110d5d5b428e105a8580d2c14a6a3d0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>