Fix tst_combobox::test_down() when popupType is set to PopupWindow
When popupType is set to PopupWindow, the geometry changes take longer
to complete. For this reason a tryCompare should be used when checking
for successful geometry changes.

The current implementation of these tests don't test for different
popupTypes, however, it is planned to do so in the future. This change
is in preparation for that.

Task-number: QTBUG-126553
Change-Id: I1204e0b4b6bdd4a7f775dcd96c3ae6765786082b
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 876131de61e316f2570d41cb16eacc9c28b2d6d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>