QQuickControl: Simplify calcLocale
There does not seem to be a reason to check arbitrary locale properties
of any item we encounter while traversing upwards; checking the locale
property of any QQuickControl derived ancestor and of the
ApplicationWindow (given there is one) ought to be enough.

This has the benefit that we don't attempt to create dynamic properties
when we encounter a type with a dynamic meta-object.

Change-Id: Ie485a2201d6bfa1ee996145beb4f9710fd804018
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>