Introduce DocumentSymbolUtils::assembleSymbolsForQmlFile
This commit introduces a main
DocumentSymbolUtils::assembleSymbolsForQmlFile function, which
employes a DocumentSymbolVisitor (which will walk through the
relevant DomItem-s, assembling DocumentSymbols along the way)
alongside some other helping functions.

Design decisions
On the top/user level we care about the "derivation"/assembling
of QList<DocumentSymbol> from  DomItem, representing QmlFile.
For that on the "high level" there is no need to manage any state,
hence free DocumentSymbolUtils::assembleSymbolsForQmlFile() is exposed.
The "assembling" function is passed to it as a dependency injection
propagated to the DocumentSymbolVisitor facilitating testing.
However because the assembling process itself involves
non trivial visiting of the DomTree, DocumentSymbolVisitor has been introduced
as "implementation details", hence it's not exposed. Another reason
for limiting the visibility was to use that entity only where it's used,
allowing to save some time on refining its API and testing,
concentrating more on the high-level functionality.

The initial aspiration behind DocumentSymbolVisitor
was to separate visiting logic from the "converting" or assembling logic.
However there are still some nuances to that.
DocumentSymbol structure mostly consists of fields
that could be derived from the DomItem obj, using DOM API.
However there is also a field `children`, which is unfortunately
not that easy to assemble using only DomItem::visitTree API.
For that some stack management is needed, which is modified
in both Opening and ClosingVisitor of DomItem::visitTree

Notes:
1. for the easier testability, it could have been better to
introduce another free function assembleSymbolsFor(DomItem),
however this time I decided to keep the interface minimal.
2. Return value of buildSymbolOrReturnChildren is SymbolsList and not
variant<Symbol,SymbolsList>, for example, is to keep the interface simpler.
Moreover, conceptually, when working with "raw" "lists" we return pointer to
the head, meaning that return type is not affected by the size of the "list".

Filtering
Unfortunately, only Fields filtering is not enough.
For example, QmlComponent stores QmlObject-s in a QList, which are
wrapped in Dom::List, however this and similar wrappers are
completely irrelevant for the DocumentSymbol and Outline, hence
they need to be filtered out based on the DomType. Another
example, is filtering out properties, which are bound at the
definition, because they will be represented by the Binding.

^^^All mentioned above could indeed be done inside the visitor itself,
however I decided to stick with the following separation of concerns:
DocumentSymbolVisitor is "hierarchy"-preserving wrapper over
DomItem::visitTree, hence fields-filtering is also part of it.
Everything what relates to the DomItem -> DocumentSymbol is
part of DocumentSymbolUtils:: namespace.

The drawback of this is there are 2 places of filtering:
1. inside the visitor (fieldsfilter)
2. inside DocumentSymbolUtils::buildSymbolOrReturnChildren

The advantages of this is
1. DocumentSymbolVisitor API is rather
minimal and its main responsibility is indeed to visit all relevant
parts of the DomTree taking care of the hierarchical order.
2. It is not that difficult to cover all relevant functionality
using dependency injection principle

Notes:
1. FieldsFilter could have been and tested explicitly as well, however
for the sake of time, it was disregarded this time
2. Inside the test, in the "af" it's possible to also add verification
of the parameters, however again, it was disregarded this time

Task-number: QTBUG-120002
Change-Id: Idd6c45e90058111b4102692c8b71d46edf7f82c6
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>